+++ CHANGELOG.md
@@ -85,7 +85,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Enhanced intellisense that filters results based on whether you typing an object proeprty/method or not.
  - switched languageserver to use [brighterscript](https://github.com/RokuCommunity/brighterscript). This is the first step towards support the BrighterScript language.
 ### Fixed
- - bug that would not register new files until a vscode restart. 
+ - bug that would not register new files until a vscode restart.
+
+
+
+## [1.22.0] - 2020-02-25
+### Added
+ - source map support during debugging. See the sourcemap section of the README for more information.
+ - support for `sourceDirs` for component libraries
+ - ability to set any default launch config settings in user/workspace settings under `brightscript.debug.launchConfigValueHere`. There is not full validation for these yet, but the logic is already in place to use them.
+### Changed
+ - setting `brightscript.rokuAdvancedLayoutEditor.trackerTaskFileLocation` has been depricated and replaced with `brightscript.debug.raleTrackerTaskFileLocation`